pyproject.toml
[project]
name = "chat-with-videos"
version = "0.1.0"
description = "Chat with your video - semantic search backend for lecture videos"
requires-python = ">=3.10"
dependencies = [
    # LanceDB + Lance (for blob API)
    "lancedb>=0.30.0",
    "pylance==2.0.1",
    "polars>=1.0.0",

    # AWS
    "boto3>=1.34.0",

    # Video processing
    "yt-dlp>=2024.1.0",
    "av>=12.0.0",
    "pillow>=10.0.0",

    # Local text embeddings (CPU-friendly)
    "sentence-transformers>=2.5.0",

    # Config
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # API server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",

    # Agents
    "pydantic-ai>=0.0.12",
    "openai>=1.12.0",

    # Async HTTP
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
